Move concept class to buisness logic package
