+++ src/main/java/com/example/demo/models/businesslogic/Concept.java
@@ -1,4 +1,4 @@
-package com.example.demo.models.entities;
+package com.example.demo.models.businesslogic;
 
 import lombok.Value;
