+++ src/main/java/com/example/demo/models/businesslogic/ExerciseForm.java
@@ -1,10 +1,8 @@
 package com.example.demo.models.businesslogic;
 
-import com.example.demo.models.entities.Concept;
 import com.example.demo.models.entities.Exercise;
 
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
